+++ crates/sml-statics/src/dec.rs
@@ -443,7 +443,7 @@ fn expansive(cx: &Cx, ars: &sml_hir::Arenas, exp: sml_hir::ExpIdx) -> bool {
     sml_hir::Exp::App(func, argument) => {
       !constructor(cx, ars, *func) || expansive(cx, ars, *argument)
     }
-    sml_hir::Exp::Typed(exp, _) => expansive(cx, ars, *exp),
+    sml_hir::Exp::Typed(exp, _, _) => expansive(cx, ars, *exp),
   }
 }
 
@@ -463,7 +463,7 @@ fn constructor(cx: &Cx, ars: &sml_hir::Arenas, exp: sml_hir::ExpIdx) -> bool {
     | sml_hir::Exp::Raise(_)
     | sml_hir::Exp::Fn(_, _) => false,
     sml_hir::Exp::Record(rows) => rows.iter().any(|&(_, exp)| constructor(cx, ars, exp)),
-    sml_hir::Exp::Typed(exp, _) => constructor(cx, ars, *exp),
+    sml_hir::Exp::Typed(exp, _, _) => constructor(cx, ars, *exp),
     sml_hir::Exp::Path(path) => {
       if path.prefix().is_empty() && path.last().as_str() == "ref" {
         return false;
@@ -513,6 +513,6 @@ fn maybe_fn(ar: &sml_hir::ExpArena, exp: sml_hir::ExpIdx) -> bool {
     | sml_hir::Exp::App(_, _)
     | sml_hir::Exp::Handle(_, _)
     | sml_hir::Exp::Raise(_) => false,
-    sml_hir::Exp::Typed(exp, _) => maybe_fn(ar, *exp),
+    sml_hir::Exp::Typed(exp, _, _) => maybe_fn(ar, *exp),
   }
 }