+++ crates/sml-statics/src/exp.rs
@@ -143,7 +143,7 @@ fn get(st: &mut St<'_>, cfg: Cfg, cx: &Cx, ars: &sml_hir::Arenas, exp: sml_hir::
       st.syms_tys.tys.fun(param, res)
     }
     // @def(9)
-    sml_hir::Exp::Typed(inner, want) => {
+    sml_hir::Exp::Typed(inner, want, _) => {
       let got = get(st, cfg, cx, ars, *inner);
       let want = ty::get(st, cx, ars, ty::Mode::Regular, *want);
       unify(st, exp.into(), want, got);
@@ -268,7 +268,7 @@ pub(crate) fn maybe_effectful(ars: &sml_hir::Arenas, exp: sml_hir::ExpIdx) -> bo
   match &ars.exp[exp] {
     sml_hir::Exp::SCon(_) | sml_hir::Exp::Path(_) | sml_hir::Exp::Fn(_, _) => false,
     sml_hir::Exp::Record(rows) => rows.iter().any(|&(_, exp)| maybe_effectful(ars, exp)),
-    sml_hir::Exp::Typed(exp, _) => maybe_effectful(ars, *exp),
+    sml_hir::Exp::Typed(exp, _, _) => maybe_effectful(ars, *exp),
     sml_hir::Exp::Let(dec, exp) => dec::maybe_effectful(ars, dec) || maybe_effectful(ars, *exp),
     sml_hir::Exp::Hole
     | sml_hir::Exp::App(_, _)